+++ examples/react/src/main.tsx
@@ -5,6 +5,8 @@ import { a } from './test';
 
 declare const BTN: string;
 declare const IRRELEVANT_ESCAPE_ENV: string;
+console.log(process.env.FARM_BASE_TEST);
+
 export function Main() {
   console.log(a);