+++ packages/cli/src/index.ts
@@ -54,6 +54,7 @@ cli
     ) => {
       const resolveOptions = resolveCommandOptions(options);
       const configPath = getConfigPath(options.config);
+
       const defaultOptions = {
         compilation: {
           root,
@@ -65,7 +66,7 @@ cli
         configPath
       };
 
-      const { start } = await resolveCore(configPath);
+      const { start } = await resolveCore();
 
       try {
         await start(defaultOptions);
@@ -110,7 +111,7 @@ cli
       },
       configPath
     };
-    const { build } = await resolveCore(configPath);
+    const { build } = await resolveCore();
 
     try {
       build(defaultOptions);
@@ -139,7 +140,7 @@ cli
       },
       configPath
     };
-    const { watch } = await resolveCore(configPath);
+    const { watch } = await resolveCore();
 
     try {
       watch(defaultOptions);
@@ -163,7 +164,7 @@ cli
       server: resolveOptions,
       configPath
     };
-    const { preview } = await resolveCore(configPath);
+    const { preview } = await resolveCore();
 
     try {
       preview(defaultOptions);
@@ -203,7 +204,7 @@ cli.version(version);
 
 cli.parse();
 
-export function resolveCore(cwd: string): Promise<{
+export function resolveCore(cwd: string = process.cwd()): Promise<{
   start: typeof start;
   build: typeof build;
   watch: typeof watch;