+++ packages/core/binding/index.d.ts
@@ -210,6 +210,7 @@ export interface Config {
     lazyCompilation?: boolean;
     treeShaking?: boolean;
     minify?: boolean;
+    record?: boolean;
     presetEnv?:
       | boolean
       | {
@@ -220,7 +221,7 @@ export interface Config {
           assumptions?: any;
         };
   };
-  jsPlugins?: object[];
+  jsPlugins?: JsPlugin[];
   // [rustPluginFilePath, jsonStringifiedOptions]
   rustPlugins?: [string, string][];
 }