+++ packages/core/src/config/schema.ts
@@ -129,6 +129,7 @@ const compilationConfigSchema = z
     lazyCompilation: z.boolean().optional(),
     treeShaking: z.boolean().optional(),
     minify: z.boolean().optional(),
+    record: z.boolean().optional(),
     presetEnv: z
       .union([
         z.boolean(),
@@ -167,6 +168,7 @@ const FarmConfigSchema = z
     root: z.string().optional(),
     base: z.string().optional(),
     clearScreen: z.boolean().optional(),
+    configPath: z.string().optional(),
     envDir: z.string().optional(),
     envPrefix: z.union([z.string(), z.array(z.string())]).optional(),
     publicDir: z.string().optional(),