+++ packages/core/src/watcher/index.ts
@@ -1,9 +1,22 @@
+import { basename, relative } from 'node:path';
+import { createRequire } from 'node:module';
+import debounce from 'lodash.debounce';
+import chalk from 'chalk';
+
 import { Compiler } from '../compiler/index.js';
 import { DevServer } from '../server/index.js';
 import { Config, JsFileWatcher } from '../../binding/index.js';
-import { compilerHandler, DefaultLogger } from '../utils/index.js';
-import debounce from 'lodash.debounce';
-import { DEFAULT_HMR_OPTIONS } from '../index.js';
+import { compilerHandler, DefaultLogger, clearScreen } from '../utils/index.js';
+import {
+  DEFAULT_HMR_OPTIONS,
+  JsPlugin,
+  normalizeUserCompilationConfig,
+  resolveUserConfig
+} from '../index.js';
+import { __FARM_GLOBAL__ } from '../config/_global.js';
+
+import type { UserConfig } from '../config/index.js';
+import { setProcessEnv } from '../config/env.js';
 
 interface ImplFileWatcher {
   watch(): Promise<void>;
@@ -17,7 +30,7 @@ export class FileWatcher implements ImplFileWatcher {
 
   constructor(
     public serverOrCompiler: DevServer | Compiler,
-    public options?: Config
+    public options?: Config & UserConfig
   ) {
     this._root = options.config.root;
     this._awaitWriteFinish = DEFAULT_HMR_OPTIONS.watchOptions.awaitWriteFinish;
@@ -42,6 +55,38 @@ export class FileWatcher implements ImplFileWatcher {
     );
 
     let handlePathChange = async (path: string): Promise<void> => {
+      const fileName = basename(path);
+      const isEnv = fileName === '.env' || fileName.startsWith('.env.');
+      const isConfig = path === this.options.inlineConfig.configPath;
+      // TODO configFileDependencies e.g: isDependencies = ["./farm.config.ts"]
+      if (isEnv || isConfig) {
+        clearScreen();
+        __FARM_GLOBAL__.__FARM_RESTART_DEV_SERVER__ = false;
+        this._logger.info(
+          `restarting server due to ${chalk.green(
+            relative(process.cwd(), path)
+          )} change`
+        );
+        if (this.serverOrCompiler instanceof DevServer) {
+          await this.serverOrCompiler.close();
+        }
+        const config: UserConfig = await resolveUserConfig(
+          this.options.inlineConfig,
+          this._logger
+        );
+        const normalizedConfig = await normalizeUserCompilationConfig(config);
+        setProcessEnv(normalizedConfig.config.mode);
+        const compiler = new Compiler(normalizedConfig);
+        const devServer = new DevServer(compiler, this._logger, config);
+        this.serverOrCompiler = devServer;
+        await devServer.listen();
+        if (normalizedConfig.config.mode === 'development') {
+          normalizedConfig.jsPlugins.forEach((plugin: JsPlugin) =>
+            plugin.configDevServer?.(devServer)
+          );
+        }
+        return;
+      }
       try {
         if (this.serverOrCompiler instanceof DevServer) {
           await this.serverOrCompiler.hmrEngine.hmrUpdate(path);
@@ -75,10 +120,6 @@ export class FileWatcher implements ImplFileWatcher {
     ]);
 
     if (this.serverOrCompiler instanceof DevServer) {
-      // chokidar.watch(
-      //   compiler.resolvedModulePaths(this._root),
-      //   this.serverOrCompiler.config.hmr.watchOptions
-      // );
       this.serverOrCompiler.hmrEngine?.onUpdateFinish((updateResult) => {
         const added = [
           ...updateResult.added,
@@ -90,30 +131,9 @@ export class FileWatcher implements ImplFileWatcher {
           );
           return resolvedPath;
         });
-
         this._watcher.watch(added);
-
-        // const removed = updateResult.removed.map((removedModule) => {
-        //   const resolvedPath = compiler.transformModulePath(
-        //     this._root,
-        //     removedModule
-        //   );
-        //   return resolvedPath;
-        // });
-
-        // this._watcher.unwatch(removed);
       });
     }
-
-    if (this.serverOrCompiler instanceof Compiler) {
-      // const watcherOptions = this.resolvedWatcherOptions();
-      // this._watcher = chokidar.watch(
-      //   compiler.resolvedModulePaths(this._root),
-      //   watcherOptions as ChokidarFileWatcherOptions
-      // );
-    }
-
-    // this._watcher.on('change', );
   }
 
   private getCompilerFromServerOrCompiler(
@@ -123,32 +143,17 @@ export class FileWatcher implements ImplFileWatcher {
       ? serverOrCompiler.getCompiler()
       : serverOrCompiler;
   }
-
-  // private resolvedWatcherOptions() {
-  //   const { watch: watcherOptions, output } = this.options.config;
-  //   const userWatcherOptions = isObject(watcherOptions) ? watcherOptions : {};
-  //   const { ignored = [] } = userWatcherOptions as ChokidarFileWatcherOptions;
-  //   const resolveWatcherOptions = {
-  //     ignoreInitial: true,
-  //     ignorePermissionErrors: true,
-  //     ...watcherOptions,
-  //     ignored: [
-  //       '**/{.git,node_modules}/**',
-  //       output?.path,
-  //       ...(Array.isArray(ignored) ? ignored : [ignored])
-  //     ]
-  //   };
-  //   // TODO other logger info
-  //   this._logger.info(`Watching for changes`);
-  //   this._logger.info(
-  //     `Ignoring changes in ${resolveWatcherOptions.ignored
-  //       .map((v: string | RegExp) => '"' + v + '"')
-  //       .join(' | ')}`
-  //   );
-  //   return resolveWatcherOptions;
-  // }
 }
 
 export async function restartServer(server: DevServer) {
   await server.close();
+}
+
+export function sleep(ms: number) {
+  return new Promise((resolve) => setTimeout(resolve, ms));
+}
+
+export function clearModuleCache(modulePath: string) {
+  const _require = createRequire(import.meta.url);
+  delete _require.cache[_require.resolve(modulePath)];
 }