+++ packages/core/tests/js-plugins.spec.ts
@@ -138,10 +138,18 @@ test('Js Plugin Execution - transform', async () => {
 test('Js Plugin Execution - full', async () => {
   const root = getJsPluginsFixturesDir();
   const resolvedPath = path.join(root, 'resolved.ts');
+  let builsStartEexcuted = false;
+  let buildEndEexcuted = false;
+
   const compiler = await getCompiler('full', [
     {
       name: 'test-full',
       priority: 1000,
+      buildStart: {
+        executor: async () => {
+          builsStartEexcuted = true;
+        }
+      },
       resolve: {
         filters: {
           sources: ['.*'],
@@ -182,12 +190,20 @@ test('Js Plugin Execution - full', async () => {
             moduleType: 'ts'
           };
         }
+      },
+      buildEnd: {
+        executor: async () => {
+          buildEndEexcuted = true;
+        }
       }
     }
   ]);
 
   await compiler.compile();
   await compiler.writeResourcesToDisk();
+
+  expect(builsStartEexcuted).toBe(true);
+  expect(buildEndEexcuted).toBe(true);
 
   const outputFilePath = getOutputFilePath('full');