+++ packages/create-farm/index.ts
@@ -214,7 +214,7 @@ function copyDir(srcDir: string, destDir: string) {
 }
 
 function welcome() {
-  console.log(chalk.magenta(`\n⚡ Welcome To Create Farm Project! \n`));
+  console.log(chalk.magenta(`\n⚡ Welcome To Farm ! \n`));
 }
 
 createFarm();