+++ packages/cli/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @farmfe/cli
 
+## 0.5.7
+
+### Patch Changes
+
+- 115b34d: chore: remove unless code
+
+## 0.5.6
+
+### Patch Changes
+
+- e90044c: fix: remove server unless options
+
 ## 0.5.5
 
 ### Patch Changes