+++ packages/core/CHANGELOG.md
@@ -1,5 +1,35 @@
 # @farmfe/core
 
+## 0.12.8
+
+### Patch Changes
+
+- 750ed61: preview command publicPath error
+
+## 0.12.7
+
+### Patch Changes
+
+- baec8bf: feat: add restart-server fn with update create-farm
+
+## 0.12.6
+
+### Patch Changes
+
+- 407e3e2: Support more js plugin hooks
+
+## 0.12.5
+
+### Patch Changes
+
+- 2d8635b: fix static assets loading issue when public path is empty
+
+## 0.12.4
+
+### Patch Changes
+
+- b5f6e88: Fix that publicPath does not work for css and static assets
+
 ## 0.12.3
 
 ### Patch Changes